+++ src/main/java/com/soorya/webscrapers/News.java
@@ -7,21 +7,24 @@ public class News
     String authorName;
     String content;
     String tags;
+    String url;
 
-    News(String timestamp,String title,String author,String content,String tags)
+    News(String timestamp, String title, String author, String content, String tags, String url)
     {
         this.timestamp = timestamp;
         this.title = title;
         this.authorName = author;
         this.content = content;
         this.tags = tags;
+        this.url = url;
     }
 
-    News(String timestamp,String title,String author,String content)
+    News(String timestamp,String title,String author,String content,String url)
     {
         this.timestamp = timestamp;
         this.title = title;
         this.authorName = author;
         this.content = content;
+        this.url = url;
     }
 }